Redpanda/requirements.txt
kafka-python
Faker
python-dotenv 